Changed Guide To Match Newest Versions
All of the guides I was finding online must have been for older versions of the rules and the modsecurity module. None of the guides provided me with a beginning to end workflow that resulted in a working configuration. This is a condensed versions of the steps I took to get a working configuration using Ubuntu 16.10, the latest Ubuntu repo version of modsecurity and the most recent OWASP rule set available through github.